requirements.txt
beautifulsoup4==4.9.3
docx==0.2.4
faiss-cpu
fastapi==0.110.3
fpdf==1.7.2
langchain==0.2.11
langchain_community==0.2.6
langchain_core==0.2.24
langchain_groq==0.1.6
langchain_huggingface==0.0.3
langchain_openai==0.1.16
langchain_text_splitters==0.2.2
langgraph==0.1.15
neo4j==5.23.1
pydantic==1.10.13
pyjwt==2.9.0
PyMuPDF
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
python_docx==1.1.2
requests==2.32.3
selenium==4.23.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.25.0
youtube-transcript-api==0.6.2
